requirements.txt
adafruit-circuitpython-ads1x15
RPi.GPIO
paho-mqtt
sanic